extract subviews to files and cleanup
