+++ WooCommerce/Classes/POS/Presentation/Settings/PointOfSaleSettingsStoreDetailView.swift
@@ -1,0 +1,16 @@
+import SwiftUI
+
+struct PointOfSaleSettingsStoreDetailView: View {
+    var body: some View {
+        NavigationStack {
+            VStack(alignment: .leading) {
+                Text("Store Settings")
+                    .font(.title2)
+                Text("Store-related configuration")
+                    .font(.caption)
+                    .foregroundStyle(.secondary)
+            }
+            .padding()
+        }
+    }
+}
